requirements.txt
requests==2.31.0
load-dotenv==0.1.0